+++ crates/ryo3-core/src/lib.rs
@@ -7,6 +7,6 @@ pub mod types;
 pub use pystring::{pystring_ascii_new, pystring_fast_new};
 
 pub use errors::FeatureNotEnabledError;
-pub use map_pyerr::map_py_value_err;
-pub use pymutex::{PyLock, PyMutex, map_poison_error};
+pub use map_pyerr::{map_py_overflow_err, map_py_runtime_err, map_py_value_err};
+pub use pymutex::{PyLock, RyMutex, map_poison_error};
 pub use pyparse::{PyFromStr, PyParse};