+++ crates/ryo3-twox-hash/src/xxhash3_64.rs
@@ -4,7 +4,7 @@ use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
 use pyo3::types::PyString;
 use pyo3::{Bound, PyResult, intern};
-use ryo3_core::PyMutex;
+use ryo3_core::RyMutex;
 use std::hash::Hasher;
 use twox_hash::XxHash3_64;
 
@@ -12,7 +12,7 @@ use twox_hash::XxHash3_64;
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3.xxhash"))]
 pub struct PyXxHash3_64 {
     seed: u64,
-    hasher: PyMutex<XxHash3_64, true>,
+    hasher: RyMutex<XxHash3_64, true>,
 }
 
 #[pymethods]