+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## v0.0.75 [unrleased]
+
+- `ryo3-jiff`
+  - Consolidate `map_py_err_*` functions into `ryo3_core`
+- `ryo3-core`
+  - renamed `PyMutex` to `RyMutex` to avoid confusion with `pyo3`'s `PyMutex`
+- `ryo3-fspath`
+  - Use `RyMutex` instead of `parking_lot::Mutex`
+- `ryo3-glob`
+  - Use `RyMutex` instead of `parking_lot::Mutex`
+
+---
+
 ## v0.0.74 [2025-12-16]
 
 - updating pyo3 to use `skip_from_py_object` where applicable (this will be a